Refactor toolbar buttons into modular components
