+++ src/components/StudyPane/Toolbar/buttons/index.ts
@@ -1,0 +1,10 @@
+export { default as ToolTip } from './ToolTip';
+export { default as UndoBtn } from './UndoBtn';
+export { default as RedoBtn } from './RedoBtn';
+export { default as ColorActionBtn } from './ColorActionBtn';
+export { default as ClearFormatBtn } from './ClearFormatBtn';
+export { default as ClearAllFormatBtn } from './ClearAllFormatBtn';
+export { default as UniformWidthBtn } from './UniformWidthBtn';
+export { default as IndentBtn } from './IndentBtn';
+export { default as StructureUpdateBtn } from './StructureUpdateBtn';
+export { default as StudyBtn } from './StudyBtn';
